+++ perl/bin/download_generate_bberg_ref_files.pl
@@ -30,6 +30,8 @@ use autodie qw(:all);
 
 use File::Path qw( remove_tree make_path );
 use File::Spec;
+use File::Copy;
+use File::Basename;
 use Getopt::Long;
 use File::Copy qw(move);
 use Pod::Usage qw(pod2usage);
@@ -148,7 +150,7 @@ sub create_one_k_genomes_files{
                 #Open relevant legend file
                 #Account for X having multiple sections in the legend files.
                 #NB this will break if not human...
-        if($THOUSAND_GENOMES_FILE_ORDER[$i] == "X"){ #If we're on chromosome X
+        if($THOUSAND_GENOMES_FILE_ORDER[$i] eq "X"){ #If we're on chromosome X
           my $j=$i;
           #use $j to iterate to the end of the list so we include all par and non par chr x positions
           for(my $j=$i; $j<scalar(@$legs); $j++){
@@ -280,7 +282,14 @@ sub download_unpack_files{
     my ($opts) = @_;
     my $impute_download = $opts->{'u'}.sprintf($IMPUTE_TGZ_PATTERN,$DOWNLOAD_VERSION);
     my $imputetgz = File::Spec->catfile($opts->{'tmp'}, sprintf($IMPUTE_TGZ_PATTERN,$DOWNLOAD_VERSION));
-    download_file($impute_download,$imputetgz,$opts->{'c'}) unless(-e $imputetgz.'.dl_success');
+    if ( defined $opts->{'a'} ) {
+      my $afile = $opts->{'a'};
+      die "\nIMPUTE archive $afile not found\n" unless ( -e $afile  );
+      die "\nFile $afile is not a tar archive\n" unless ( $afile =~ m/.tar.gz$/ or $afile =~ m/.tgz$/);
+      copy($afile,$imputetgz);
+    } else {
+      download_file($impute_download,$imputetgz,$opts->{'c'}) unless(-e $imputetgz.'.dl_success');
+    }
     unpack_file($imputetgz,$opts->{'tmp'},'tgz');
     return;
 }
@@ -294,6 +303,7 @@ sub unpack_file{
 
 sub download_file{
     my ($url,$file,$use_curl) = @_;
+    print "downloading\n";
         if ($use_curl) {
           my $output = `curl --location $url > $file`;
         } else {
@@ -326,6 +336,7 @@ sub setup{
                     'v|version' => \$opts{'v'},
                     'c|use-curl' => \$opts{'c'},
                     'o|out-dir=s' => \$opts{'o'},
+                    'a|archive=s' => \$opts{'a'},
                     'd|download-version=s' => \$opts{'d'},
           'u|url=s' => \$opts{'u'},
           '<>' => sub{push(@random_args,shift(@_));},
@@ -343,9 +354,12 @@ sub setup{
   # make outdir absolute
   $opts{'o'} = File::Spec->rel2abs( $opts{'o'} );
 
+    
     #Ensure download and other directories exist, if not create it.
     my $tmpdir = File::Spec->catdir($opts{'o'}, 'tmp');
     make_path($tmpdir) unless(-d $tmpdir);
+
+    die "\nSpecified impute archive doesn't exist\n" if ( defined $opts{'a'} &&  (not -e $opts{'a'}) );    
 
     my $imputedir = File::Spec->catdir($opts{'o'}, 'impute');
     make_path($imputedir) unless(-d $imputedir);
@@ -381,6 +395,7 @@ download_generate_bberg_ref_files.pl [options]
     Optional parameters:
       -url                      -u  URL to download battenberg impute reference data from [see docs for default url]
       -use-curl                 -c  Use curl for the download
+      -archive                  -a  Specify pre-downloaded IMPUTE archive
       -download-version         -d  Download version (part of download name) default: [v3]
 
     Other:
@@ -406,6 +421,11 @@ URL to download impute files from. Default is [https://mathgen.stats.ox.ac.uk/im
 Files downloaded are:
 ALL_1000G_phase1integrated_v3_annotated_legends.tgz
 ALL_1000G_phase1integrated_v3_impute.tgz
+
+=item B<-archive>
+
+Pre-downloaded IMPUTE archive
+
 
 =item B<-download-version>